+++ packages/gov-ui-kit/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
+-   Rename default core context values to use `guk` prefix
 -   Update `<DefinitionListItem />` and `<Breadcrumbs />` core components to truncate long strings
 -   Fix `<InputNumber >` core component to trigger `onChange` callback on + / - button click